requirements.txt
Flask
Pillow
opencv-python
pytesseract
numpy
